// src/app/modules/tecnologia-informacao/estoque/produtos/models/produto.ts
export interface ITecnologiaInformacaoEstoqueProduto {
  ID_TEIN_ESTO_PROD: number;
  ID_TEIN_ESTO_PRMA: number;
  ID_TEIN_ESTO_PRTP: number;
  ID_TEIN_ESTO_PRMO: number;
  NM_PROD: string;
  NM_TIPO: string;
  NM_MODE: string;
  DS_UNID_MEDI: string;
  TT_PESO_UNIT: number;
  IN_CODI_SERI: string;
  IN_STAT: string;
  DT_INCL: string;
  IN_PAGI: string;
  DS_OBSE: string;
  NM_USUA_CADA: string;
  NM_MATR_CADA: number;
}